% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{allgraphs}
\alias{allgraphs}
\title{Pipeline function}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
allgraphs(datafile, outdir, mapfile, sampdepth = NULL, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{mapfile}{full path to map file}

\item{sampdepth}{sampling depth.  see details.}

\item{...}{other parameters to pass to \link{readindata}}
}
\value{
graphs are saved to outdir.  See [user doc](../doc/user_doc.md).
}
\description{
Make all 4 types of graphs
}
\details{
If sampdepth is NULL, then the sampling depth is set to the size of the
smallest sample larger than 0.2*median sample size.  Otherwise, it is set to the
size of the smallest sample larger than sampdepth.

This value is used to remove samples before for alpha diversity and PCoA plots.
Also, to rarefy OTU table for the alpha diversity and Bray-Curtis distance PCoA.
}
