% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ampvis2biom}
\alias{ampvis2biom}
\title{Convert ampvis2 object to biom}
\usage{
ampvis2biom(amp, file = NULL, id = NULL, matrix_element_type = c("int",
  "float"))
}
\arguments{
\item{amp}{ampvis2 object like from output of \code{\link[ampvis2:amp_load]{ampvis2::amp_load}}}

\item{file}{(Optional). File name to write (BIOM V1 json).}

\item{id}{(Optional). Character string. Identifier for the project.}

\item{matrix_element_type}{(Optional). Character string. Either 'int' or 'float'}
}
\value{
invisibly returns \link[biomformat]{biom} object
}
\description{
Converts ampvis2 object to biom and optionally writes file.  Uses 
\link[biomformat]{make_biom} and modifies object to make it validate with 
\url{https://biom-format.org/}.
}
