% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{dataviz16s}
\alias{dataviz16s}
\alias{dataviz16s-package}
\title{dataviz16s: A package for Nephele 16S pipeline visualization}
\description{
dataviz16s: A package for Nephele 16S pipeline visualization
}
