% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{morphheatmap}
\alias{morphheatmap}
\title{Morpheus heatmap}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
morphheatmap(mapfile, datafile, outdir, amp, sampdepth = NULL,
  rarefy = FALSE, filter_level = 0.1, taxlevel = c("seq"), ...)
}
\arguments{
\item{mapfile}{full path to mapping file}

\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{amp}{(Optional) ampvis2 object. may be specified instead of mapfile and datafile}

\item{sampdepth}{sampling depth}

\item{rarefy}{Logical. Rarefy the OTU table if sampdepth is specified.}

\item{filter_level}{minimum abundance to show in the heatmap}

\item{taxlevel}{vector of taxonomic levels to graph.  must be subset of
c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "seq").  See Details.}

\item{...}{parameters to pass to  \code{\link{readindata}}}
}
\value{
Saves heatmaps to outdir.
}
\description{
Creates heatmaps using Morpheus R API \url{https://software.broadinstitute.org/morpheus/}.  The heatmaps are made
using relative abundances.
}
\details{
For the \code{taxlevel} parameter, each level is made into a separate heatmap.  "seq" makes
 the heatmap with no collapsing of taxonomic levels.
}
\examples{
\dontrun{
 morphheatmap(mapfile="mapfile.txt", datafile="OTU_table.txt", outdir="outputs/graphs",
 sampdepth = 25000, taxlevel = c("Family", "seq"), tsvfile=TRUE)
}


}
