% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rarefactioncurve}
\alias{rarefactioncurve}
\title{Make rarefaction curve graph}
\usage{
rarefactioncurve(mapfile, datafile, outdir, amp, colors = NULL, ...)
}
\arguments{
\item{mapfile}{full path mapping file}

\item{datafile}{full path to OTU file}

\item{outdir}{full path to output directory}

\item{amp}{(Optional) ampvis2 object. may be specified instead of mapfile and datafile}

\item{colors}{(Optional) color vector - length equal to number of TreatmentGroups in mapfile}

\item{...}{parameters to pass to \code{\link{readindata}}}
}
\value{
Saves rarefaction curve plot to output directory.
}
\description{
Make rarefaction curve graph
}
