% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{shortnames}
\alias{shortnames}
\title{shortnames for SILVA taxonomy}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
shortnames(taxtable, taxa = "Species")
}
\arguments{
\item{taxtable}{taxonomy table object from amp$tax}

\item{taxa}{taxonomic level at which to retrieve the names}
}
\value{
Vector of shortened names for each seqvar/otu.
}
\description{
shortnames for SILVA taxonomy
}
