% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{adivboxplot}
\alias{adivboxplot}
\title{Alpha diversity boxplot}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
adivboxplot(mapfile, datafile, outdir, amp, sampdepth = NULL, colors = NULL,
  ...)
}
\arguments{
\item{mapfile}{full path to map file}

\item{datafile}{full path to input OTU file}

\item{outdir}{full path to output directory}

\item{amp}{ampvis2 object. may be specified instead of mapfile and datafile}

\item{sampdepth}{sampling depth}

\item{colors}{colors to use for plots}

\item{...}{other parameters to pass to \link{readindata}}
}
\value{
Save alpha diversity boxplots to outdir.
}
\description{
Plots exploding boxplot of shannon diversity and Chao species richness.  If sampling depth is NULL,
rarefies OTU table to the minimum readcount of any sample.  If this is low, then the plot will fail.
}
