% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{allgraphs}
\alias{allgraphs}
\title{Pipeline function}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
allgraphs(mapfile, datafile, outdir, sampdepth = NULL, ...)
}
\arguments{
\item{mapfile}{full path to map file}

\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{sampdepth}{sampling depth}

\item{...}{other parameters to pass to \link{readindata}}
}
\value{
graphs are saved to outdir.  See [user doc](../doc/user_doc.md).
}
\description{
Make all 4 types of graphs
}
