% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{trygraphwrapper}
\alias{trygraphwrapper}
\title{Wrapper for any graph function}
\usage{
trygraphwrapper(mapfile, datafile, outdir, FUN, logfilename = "logfile.txt",
  info = TRUE, ...)
}
\arguments{
\item{mapfile}{full path to map file}

\item{datafile}{full path input OTU file}

\item{outdir}{output directory for graphs}

\item{FUN}{function you would like to run}

\item{logfilename}{logfilename}

\item{info}{print sessionInfo to logfile}

\item{...}{parameters needed to pass to FUN}
}
\value{
Returns 0 if FUN succeeds and 1 if it returns an error.
}
\description{
This is a wrapper for any of the graph functions meant to be called using rpy2 in python.
}
\examples{

\dontrun{
trygraphwrapper("/mnt/EFS/user_uploads/job_id/inputs/mapfile.txt","/mnt/EFS/user_uploads/job_id/outputs/out.biom",
"/mnt/EFS/user_uploads/job_id/outputs/", allgraphs)
# example with no optional arguments for running allgraphs
}

\dontrun{
trygraphwrapper("/mnt/EFS/user_uploads/job_id/inputs/mapfile.txt","/mnt/EFS/user_uploads/job_id/outputs/out.biom",
"/mnt/EFS/user_uploads/job_id/outputs/", allgraphs, sampdepth = 30000)
# example with optional argument sampdepth
}

}
