% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{trygraphwrapper}
\alias{trygraphwrapper}
\title{Wrapper for any graph function}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
trygraphwrapper(mapfile, datafile, outdir, FUN, logfilename = "logfile.txt",
  info = TRUE, ...)
}
\arguments{
\item{mapfile}{full path to map file}

\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{output directory for graphs}

\item{FUN}{character string. name of function you would like to run. can be actual
function object if run from R}

\item{logfilename}{logfilename}

\item{info}{print sessionInfo to logfile}

\item{...}{parameters needed to pass to FUN}
}
\value{
Returns 0 if FUN succeeds and 1 if it returns an error.
}
\description{
This is a wrapper for any of the graph functions meant to be called using rpy2 in python.
}
\examples{

\dontrun{

# example with no optional arguments for running allgraphs
trygraphwrapper("/path/to/inputs/mapfile.txt","/path/to/outputs/out.biom",
 "/path/to/outputs/", 'allgraphs')

# example with optional argument sampdepth
trygraphwrapper("/path/to/inputs/mapfile.txt","/path/to/outputs/out.biom",
"/path/to/outputs/", 'allgraphs', sampdepth = 30000)

# example of making heatmap with optional arguments
trygraphwrapper("/path/to/inputs/mapfile.txt", "/path/to/outputs/taxa_species.biom",
"/path/to/outputs", 'morphheatmap', sampdepth = 30000, filter_level=0.01,
taxlevel=c("Family", "seq"))
}


}
